% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commDD.R
\name{commDD}
\alias{commDD}
\title{Calculate Molecular Dendrogram-based Diversity}
\usage{
commDD(mol.data, mol.dendrogram, type = c("DD", "MPD", "MNTD"))
}
\arguments{
\item{mol.data}{Data frame containing molecular data where rows represent different samples or conditions, and columns represent individual molecules. Default: mol.data.}

\item{mol.dendrogram}{The dendrogram object representing the relationships among molecules.}

\item{type}{A character vector specifying the types of diversity indices to calculate. Default: c("DD", "MPD", "MNTD").}
}
\value{
A data frame with the calculated diversity indices for each sample.
}
\description{
This function calculates various Dendrogram-based diversity indices for molecular data.
}
\seealso{
\code{\link[vegan]{decostand}}
\code{\link[picante]{pd}}, \code{\link[picante]{mpd}}, \code{\link[picante]{mntd}}
}
