% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commDis.R
\name{commDis}
\alias{commDis}
\title{Calculate dissimilarity between molecular communities}
\usage{
commDis(mol.data, method = "braycurtis", mol.dendrogram = NULL)
}
\arguments{
\item{mol.data}{Data frame containing molecular data where rows represent different samples or conditions and columns represent individual molecules. Default: mol.data}

\item{method}{Dissimilarity method to use. Options: "jaccard", "braycurtis", "unweighted_unifrac", "weighted_unifrac". Default: "braycurtis".}

\item{mol.dendrogram}{The dendrogram object representing the relationships among molecules. Required for UniFrac methods. Default: NULL}
}
\value{
Dissimilarity matrix
}
\description{
This function calculates dissimilarity metrics between molecular communities based on given method.
}
\seealso{
\code{\link[vegan]{vegdist}}
\code{\link[phyloseq]{phyloseq}}, \code{\link[phyloseq]{UniFrac}}
}
