% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commTD.R
\name{commTD}
\alias{commTD}
\title{Calculate Molecular Taxonomic Diversity}
\usage{
commTD(mol.data, type = c("Richness"))
}
\arguments{
\item{mol.data}{Data frame containing molecular data where rows represent different samples or conditions, and columns represent individual molecules. Default: mol.data.}

\item{type}{A character vector specifying the diversity indices to calculate. Options are "Richness", "Shannon", "Simpson", "Invsimpson", "PielouEven", and "Chao1". Default: Richness.}
}
\value{
A data frame containing calculated diversity indices for each sample.
}
\description{
This function calculates various taxonomic diversity indices for molecular data.
}
\seealso{
\code{\link[vegan]{decostand}}, \code{\link[vegan]{diversity}}, \code{\link[vegan]{specpool}}
}
