% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCER.R
\name{iCER}
\alias{iCER}
\title{The Indicator of Compositional-Level Thermal Responses (iCER)}
\usage{
iCER(
  data_ra,
  envi,
  prop = 0.8,
  Temp_num,
  end_n = 10,
  cutoff = 0.3,
  vars_list = "Temperature",
  significant_only = FALSE,
  p_value_threshold = 0.05
)
}
\arguments{
\item{data_ra}{A data frame containing the relative abundance of molecules, where rows represent samples and columns represent different molecules.}

\item{envi}{A data frame containing environmental variables associated with each sample. The rows should correspond to the samples in \code{data_ra}, and it should include the variable(s) specified in \code{vars_list}.}

\item{prop}{The proportion of the dataset to be used for MER (Molecule-specific thermal responses) calculation. This controls how much of the original dataset is sampled for analysis. Default: 0.8.}

\item{Temp_num}{The number of temperature levels randomly selected in each permutation.}

\item{end_n}{Ensures that each sample appears at least this many times in the iCER datasets during random data partitioning. Increasing this value will result in a greater number of permutations. Default: 10.}

\item{cutoff}{The threshold for retaining molecules observed in a specified proportion of the samples within the MER dataset. Molecules present in less than this proportion of samples will be excluded from the MER calculation. Default: 0.3.}

\item{vars_list}{The name or list of environmental variables (such as temperature) to be used in the correlation analysis. Default: 'Temperature'.}

\item{significant_only}{A boolean indicating whether to calculate iCER using only molecules with significant MERs. If set to TRUE, only molecules with significant correlations (based on \code{p_value_threshold}) are included. Default: FALSE.}

\item{p_value_threshold}{The significance level for determining whether a molecule’s correlation is considered significant. Default: 0.05}
}
\value{
A list containing the following components:
\itemize{
\item \code{MER}: A data frame of MER values for molecules.
\item \code{CER_all_out}: A data frame of iCER values calculated from all molecules (if \code{significant_only = FALSE}).
\item \code{CER_sig_out}: A data frame of iCER values calculated from molecules with significant MERs (if \code{significant_only = TRUE}).
\item \code{CER_mean}: A data frame of the mean iCER value for each sample.
}
}
\description{
The iCER function calculates the Indicator of Compositional-Level Thermal Responses (iCER) based on environmental data (\code{envi}) and molecular relative abundance data (\code{data_ra}).
}
\examples{
\dontrun{
if(interactive()){
 # Example of calculating iCER using all molecules
 result_all <- iCER(data_ra, envi, significant_only = FALSE)
 
 # Example of calculating iCER using only significant molecules
 result_sig <- iCER(data_ra, envi, significant_only = TRUE, p_value_threshold = 0.05)
 }
}
}
\seealso{
\code{\link[FD]{functcomp}}
}
