% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iDME.R
\name{iDME}
\alias{iDME}
\title{The Indicator of Molecular Dark Matter Effects (iDME)}
\usage{
iDME(
  mol.data,
  mol.dark.matter,
  occu.rate = 0.3,
  bootstrap = 100,
  Network.size = 400,
  sparcc.R = NULL,
  sparcc.R.threshold = 0.3
)
}
\arguments{
\item{mol.data}{Data frame containing molecular data where rows represent different samples or conditions, and columns represent individual molecules. Default: mol.data.}

\item{mol.dark.matter}{Data frame containing molecular dark matter where rows represent different samples or conditions, and columns represent individual unclassified molecules. Default: mol.dark.matter.}

\item{occu.rate}{The threshold for retaining DOM molecules observed in a specified proportion of the total samples. Default is 0.3.}

\item{bootstrap}{The number of 'KK' and 'DK' networks to be randomly generated for each sample; defaults to 100 (more might be better, less probably not).}

\item{Network.size}{The size of 'KK' and 'DK' networks. Default: 400.}

\item{sparcc.R}{An optional precomputed SparCC correlation matrix. If NULL, the function will compute it. Default is NULL.}

\item{sparcc.R.threshold}{A numeric value specifying the threshold for the absolute value of the SparCC correlation matrix to filter out uncorrelated or weakly correlated interactions. Correlations with an absolute value less than 0.3 (i.e., -0.3 < correlation < 0.3) will be removed. Default is 0.3.}
}
\value{
A data frame with calculated iDME, iDME.intra, and iDME.inter for each sample.
}
\description{
The function calculates the iDME by measuring the percentage change in the average value of a specified network metric between the “KK” (known-known) and “DK” (dark-known) networks.
}
\seealso{
\code{\link[vegan]{decostand}}
\code{\link[SpiecEasi]{sparcc}}
\code{\link[igraph]{graph_from_adjacency_matrix}}, \code{\link[igraph]{V}}, \code{\link[igraph]{degree}}
\code{\link[bipartite]{specieslevel}}
\code{\link[dplyr]{rename}}
}
