% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iHaf.R
\name{iHaf}
\alias{iHaf}
\title{The Indicator of Molecular Habitat Affinity (iHaf)}
\usage{
iHaf(data.ra.habi1, data.ra.habi2)
}
\arguments{
\item{data.ra.habi1}{A data frame containing the relative abundance of molecules, where rows represent samples and columns represent different molecules.}

\item{data.ra.habi2}{A data frame containing the relative abundance of molecules, where rows represent samples and columns represent different molecules.}
}
\value{
A list containing the following components:
\describe{
\item{iHaf}{A data frame of iHaf values for molecules.}
\item{iHaf_wm_habi1}{A data frame of iHaf_wm values for samples of habitat1.}
\item{iHaf_wm_habi2}{A data frame of iHaf_wm values for samples of habitat2.}
}
}
\description{
The iHaf function calculates the Indicator of Molecular Habitat Affinity (iHaf) based on molecular relative abundance data of habitat1 (data.ra.habi1) and habitat2 (data.ra.habi2).
}
\examples{
\dontrun{
iHaf(data.ra.habi1, data.ra.habi2)
}
}
