% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molGroup.R
\name{molGroup}
\alias{molGroup}
\title{Molecular Grouping Analysis}
\usage{
molGroup(
  mol.data,
  mol.trait,
  group = "Reactivity_Activity",
  HtoC_ratio,
  Peak.trans.profile = NULL,
  Trans_threshold = c(1, 10)
)
}
\arguments{
\item{mol.data}{Data frame containing molecular data where rows represent different samples or conditions, and columns represent individual molecules. Default: mol.data}

\item{mol.trait}{Data frame containing molecular trait data such as mass or other properties associated with each molecule. Default: mol.trait}

\item{group}{The grouping criteria, Default: 'Reactivity_Activity'}

\item{Trans_threshold}{The threshold for transformations, Default: c(1, 10)}
}
\value{
A data frame with grouped molecules.
}
\description{
This function groups molecular data based on specific traits.
}
