% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molTrait.R
\name{molTrait}
\alias{molTrait}
\title{Molecular Trait Analysis}
\usage{
molTrait(report)
}
\arguments{
\item{report}{A data frame containing the molecular report data generated by Formularity (Tolić et al, 2017 - Anal. Chem.)}
}
\value{
A data frame of the molecular traits.
}
\description{
This function performs a comprehensive analysis of molecular traits from a given report
}
