% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRA.R
\name{plotRA}
\alias{plotRA}
\title{Plot Relative Abundance by Group}
\usage{
plotRA(mol.data, mol.group, group_col = "group")
}
\arguments{
\item{mol.data}{Data frame containing molecular data where rows represent different samples or conditions, and columns represent individual molecules. Default: mol.data}

\item{mol.group}{Data frame containing groups for molecular data.}

\item{group_col}{Column name in mol.group indicating groups. Default: 'group'}
}
\value{
A ggplot object.
}
\description{
Plot relative abundance of molecular data by groups.
}
