% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVK.R
\name{plotVK}
\alias{plotVK}
\title{Plot Van Krevelen Diagram}
\usage{
plotVK(data, HtoC_ratio, OtoC_ratio, title = "Van Krevelen Diagram")
}
\arguments{
\item{data}{A data frame containing the H/C ratios and O/C ratios.}

\item{HtoC_ratio}{The name of the column in \code{data} containing H/C ratios.}

\item{OtoC_ratio}{The name of the column in \code{data} containing O/C ratios.}

\item{title}{Title of the plot. Default: 'Van Krevelen Diagram'}
}
\value{
A ggplot object representing the Van Krevelen diagram.
}
\description{
Function to plot a Van Krevelen diagram.
}
\details{
This function creates a scatter plot of H/C vs O/C ratios.
}
