% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commFD.R
\name{commFD}
\alias{commFD}
\title{Calculate Molecular Functional Diversity}
\usage{
commFD(mol.data, mol.trait, trait_col = "AI_Mod")
}
\arguments{
\item{mol.data}{Data frame containing molecular data where rows represent different samples or conditions, and columns represent individual molecules. Default: mol.data.}

\item{mol.trait}{Data frame containing molecular trait data such as mass or other properties associated with each molecule. Default: mol.trait.}

\item{trait.name}{The name of the trait column in the \code{mol.trait}. Default: "AI_Mod".}
}
\value{
A data frame containing functional diversity indices for each sample.
}
\description{
This function calculates functional diversity indices using molecular data and trait data.
}
\seealso{
\code{\link[vegan]{decostand}}
\code{\link[fundiversity]{fd_raoq}}
}
