% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commProc.R
\name{commProc}
\alias{commProc}
\title{Quantifying Assembly Processes of DOM assemblages}
\usage{
commProc(mol.data, mol.dendrogram, mol.group = NULL, group_col = "group")
}
\arguments{
\item{mol.data}{Data frame containing molecular data where rows represent different samples or conditions, and columns represent individual molecules. Default: mol.data.}

\item{mol.dendrogram}{The dendrogram object representing the relationships among molecules.}

\item{mol.group}{Optional data frame containing group information for molecular data. Default: NULL.}

\item{group_col}{Column name in \code{mol.group} indicating groups. This parameter is required if \code{mol.group} is provided.}
}
\value{
A list with the following elements:
\item{ratio}{A data frame showing the overall percentage of turnovers governed by each process.}
\item{result}{A matrix showing betaMNTD, Bray-Curtis dissimilarity, betaNTI, RC, and the identified process governing each turnover. Each row represents a turnover between two samples.}
}
\description{
This function quantifies the relative influences of deterministic and stochastic processes governing the assembly of DOM assemblages.
}
