% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H2.R
\name{H2}
\alias{H2}
\title{The specialization index H2' of DOM-microbe associations}
\usage{
H2(
  mol.data,
  micro.data,
  occu.rate = 0.3,
  sparcc.R = NULL,
  sparcc.R.threshold = 0.3,
  N = 100,
  Null.model = "shuffle.web"
)
}
\arguments{
\item{mol.data}{Sample/DOM matrix with samples in the rows and DOM molecules in the columns (compositional/abundant data).}

\item{micro.data}{Sample/Microbe matrix with samples in the rows and Microbial species in the columns (compositional/abundant data).}

\item{occu.rate}{The threshold for retaining bacterial species or DOM molecules observed in a specified proportion of the total samples. Default is 0.3.}

\item{sparcc.R}{An optional precomputed SparCC correlation matrix. If NULL, the function will compute it. Default: NULL}

\item{sparcc.R.threshold}{The threshold of including correlations between DOM molecules and bacterial species in bipartite networks. Default: 0.3}

\item{N}{Number of null models to be generated; defaults to 100 (more might be better, less probably not).}

\item{Null.model}{Null model type. Can be given as an integer or name: 1/"r2dtable", 2/"swap.web", 3/"vaznull", 4/"shuffle.web"; allows for partial match of names. Default: 'shuffle.web'.}
}
\value{
Returns a data.frame, which contains standardised H2', Observed H2', P-value and Network type. Standardised: Standardised H2'. It is standardised by using a null modelling approach.
Observed: Observed H2'. It ranges between 0 (complete generalization) and 1 (complete specialization). p.value: The significance of differences between observed and random H2'. Network.type: Network type. Can be given as a name: Full (full network),
Negative (negative network), Positive (positive network).
}
\description{
This function calculates the network-level specialization of all interacting trophic levels in DOM-microbe bipartite networks, including full, negative and positive networks.
}
\details{
H2' is a network-level property to describe how much the two trophic levels are interacting with each other in a bipartite network.
For example, H2' is used to quantify the specialization of DOM-microbe associations at a network level.
Specifically, elevated H2' values convey that there is a high degree of specialization between DOM and microbes.
By contrast, lower H2' values reflect a more generalized bipartite network where different DOM molecules can be used by a large range of bacterial taxa.
}
\seealso{
\code{\link[vegan]{decostand}}
\code{\link[SpiecEasi]{sparcc}}
\code{\link[bipartite]{nullmodel}}, \code{\link[bipartite]{networklevel}}
}
