% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCER.R
\name{iCER}
\alias{iCER}
\title{The Indicator of Compositional-Level Thermal Responses (iCER)}
\usage{
iCER(
  data_ra,
  envi,
  prop = 0.8,
  Temp_num,
  end_n = 10,
  cutoff = 0.3,
  vars_list = "Temperature",
  significant_only = FALSE,
  p_value_threshold = 0.05
)
}
\arguments{
\item{data_ra}{A data frame containing the relative abundance of molecules, where rows represent samples and columns represent different molecules.}

\item{envi}{A data frame containing environmental variables associated with each sample. The rows should correspond to the samples in \code{data_ra}, and it should include the variable(s) specified in \code{vars_list}.}

\item{prop}{The proportion of samples selected in each permutation for MER (molecule-specific thermal response) calculation. This determines the 'MER dataset' size and affects the stability and variability of the results. Default: 0.8.}

\item{Temp_num}{The number of temperature levels randomly selected in each permutation.}

\item{end_n}{The minimum number of times each sample must appear in the 'iCER datasets' during random data partitioning. Increasing this value will result in a greater number of permutations. Default: 10.}

\item{cutoff}{The minimum proportion of samples in the 'MER dataset' in which a molecule must appear to be retained; molecules below this cutoff are excluded from MER estimation. Default: 0.3.}

\item{vars_list}{The name of environmental variables (such as 'Temperature') to be used for MER calculation. Default: 'Temperature'.}

\item{significant_only}{Whether to compute iCER using only molecules with statistically significant MER values (p ≤ \code{p_value_threshold}). Default: FALSE.}

\item{p_value_threshold}{The p-value threshold used to assess whether a molecule’s MER is statistically significant. Default: 0.05.}
}
\value{
A list containing the following components:
\itemize{
\item \code{MER}: A data frame of MER values and p-values for molecules in each permutation.
\item \code{iCER_all_permutation}: A data frame containing iCER, weighted variance (\code{iCER.var}), standard deviation (\code{iCER.sd}), and permutation index for each sample, computed from all molecules (if \code{significant_only = False}).
\item \code{iCER_sig_permutation}: A data frame containing iCER, weighted variance (\code{iCER.var}), standard deviation (\code{iCER.sd}), and permutation index for each sample, computed from molecules with significant MERs (if \code{significant_only = TRUE}).
\item \code{iCER}: A data frame containing the aggregated iCER results for each sample across all permutations, including the mean iCER, its weighted variance (\code{iCER.var}), and standard deviation (\code{iCER.sd}).
}
}
\description{
The iCER function calculates the Indicator of Compositional-Level Thermal Responses (iCER) based on environmental data (\code{envi}) and molecular relative abundance data (\code{data_ra}).
}
\examples{
\dontrun{
if(interactive()){
 # Example of calculating iCER using all molecules
 result_all <- iCER(data_ra, envi, significant_only = FALSE)
 
 # Example of calculating iCER using only significant molecules
 result_sig <- iCER(data_ra, envi, significant_only = TRUE, p_value_threshold = 0.05)
 }
}
}
\seealso{
\code{\link[FD]{functcomp}}
\code{\link[Hmisc]{wtd.var}}
}
