% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iHaf.R
\name{iHaf}
\alias{iHaf}
\title{The Indicator of Molecular Habitat Affinity (iHaf)}
\usage{
iHaf(data.ra.habi1, data.ra.habi2)
}
\arguments{
\item{data.ra.habi1}{A data frame containing the relative abundance of molecules, where rows represent samples and columns represent different molecules.}

\item{data.ra.habi2}{A data frame containing the relative abundance of molecules, where rows represent samples and columns represent different molecules.}
}
\value{
iHaf: A data frame of iHaf values for molecules.
}
\description{
The iHaf function calculates the Indicator of Molecular Habitat Affinity (iHaf) based on molecular relative abundance data of habitat1 (data.ra.habi1) and habitat2 (data.ra.habi2).
}
\examples{
\dontrun{
if(interactive()){
 iHaf(data.ra.habi1,data.ra.habi2)
 }
}
}
