% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molTrait.R
\name{molTrait}
\alias{molTrait}
\title{Molecular Trait Analysis}
\usage{
molTrait(report, ver = "1.1.0")
}
\arguments{
\item{report}{A data frame containing the molecular report data generated by Formularity (Tolić et al, 2017, Analytical Chemistry)}

\item{ver}{The version of the ftmsRanalysis package to use. Supports "1.1.0" and "1.2.0". Default is "1.1.0".}
}
\value{
A data frame of the molecular traits
}
\description{
This function performs a comprehensive analysis of molecular traits from a given report
}
\examples{
\dontrun{
report = read.csv(file = list.files(path = ".", pattern = "*Report.csv", full.names = T))
mol.trait = molTrait(report, ver = "1.1.0")}
}
